+++ src/features/admin/evaluation-administrations/[id]/preview/preview-evaluees-header.tsx
@@ -3,7 +3,7 @@ import { PageTitle } from "@components/shared/page-title"
 export const PreviewEmployeesHeader = () => {
   return (
     <div>
-      <PageTitle>Review Employees</PageTitle>
+      <PageTitle>Review Evaluees</PageTitle>
     </div>
   )
 }
